requirements.txt
flask
gunicorn
pillow
opencv-python
tensorflow-cpu
